implemented an abstract class that allows the user to have a more direct control of how the ValueProperty binds. That class is BaseValidationBehavior. The ValidationBehavior inherits from this one but now inherits from the base, and sets everything the same way as before. The others should experience any changes. Time to test this.
